requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
celery==3.1.23
Django==1.9.6
django-extra-views==0.6.4
django-haystack==2.5.0
django-localflavor==1.3
django-oscar==1.2.1
django-oscar-accounts==0.4rc1
django-oscar-api==1.0.4
django-oscar-api-checkout==0.1.4
django-oscar-bluelight==0.5.0
django-oscar-wfrs==0.2.6
django-statsd-mozilla==0.3.16
django-tables2==1.0.7
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
elasticsearch==1.9.0
factory-boy==2.6.1
fake-factory==0.7.2
flake8==2.5.4
instrumented-soap==1.1.0
kombu==3.0.37
mccabe==0.4.0
mock==1.3.0
pbr==1.10.0
pep8==1.7.0
phonenumbers==7.7.5
Pillow==3.4.2
psycopg2==2.6.1
purl==1.3
pyflakes==1.0.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.11
requests==2.12.4
six==1.10.0
sorl-thumbnail==12.3
statsd==2.1.2
suds-jurko==0.6
Unidecode==0.4.19
urllib3==1.19.1
versiontag==1.0.3
